requirements.txt
click
diskcache
openstacksdk>=0.9.14
psutil>=5.2.0
six
stevedore
